pyproject.toml
[project]
name = "opcoord-platform"
version = "0.1.0"
description = "Baltic regional coordination tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elasticsearch==8.19.0",
    "loguru>=0.7.3",
    "pandas>=3.0.1",
    "pika>=1.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "rdflib>=7.6.0",
    "requests>=2.32.5",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "local-pypi"
url = "https://artifactory.elering.sise/artifactory/api/pypi/pssepip/simple"
#default = true

[tool.uv]
allow-insecure-host = ["artifactory.elering.sise"]
